dev_requirements.txt
black==21.9b0
coverage==6.1.1
functions-framework==3.0.0
mypy==0.910
mypy-extensions==0.4.3
pep8-naming==0.10.0
pylama-pylint==3.1.1
pylint==2.5.3
pylint-flask==0.6
pytest==6.2.5
radon==5.0.1
pytest-cov==3.0.0
types-Flask==1.1.2
types-requests==2.25.11
